+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add support for the `FENCE` instruction
 - Add support for DRAMsys5.0 co-simulation
 - Add support for atomics in L2
+- Add support for a multi-cluster MemPool
 
 ### Changes
 - Add physical feasible TeraPool configuration with SubGroup hierarchy.